add authorization password and file password to categories

+++ controllers/categoryController.js
@@ -1,4 +1,7 @@
 const { body, validationResult } = require('express-validator');
+const multer = require('multer');
+const { authorizationValidator, fileValidator, makeFileFilter } = require('../helpers/extraValidators');
+
 const Category = require('../models/category');
 const Item = require('../models/item');
 
@@ -11,6 +14,11 @@ body('description', 'Description must not be empty')
   .isLength({ min: 1 })
   .escape(),
 ];
+
+const upload = multer({
+  dest: './public/data/uploads',
+  fileFilter: makeFileFilter('file_password'),
+});
 
 // index
 exports.index = (req, res, next) => {
@@ -59,12 +67,15 @@ exports.category_create_get = (req, res, next) => {
 
 // handle category create form
 exports.category_create_post = [
+  upload.single('uploaded_file'),
   [...categoryFormValidators],
+  fileValidator,
   (req, res, next) => {
     const errors = validationResult(req);
     const category = new Category({
       name: req.body.name,
       description: req.body.description,
+      image_path: (req.file) ? req.file.filename : 'none',
     });
 
     if (!errors.isEmpty()) {
@@ -109,30 +120,34 @@ exports.category_delete_get = (req, res, next) => {
 };
 
 // handle category delete form
-exports.category_delete_post = (req, res, next) => {
-  Item.find({ category: req.params.id })
-    .then((items) => {
-      if (items.length > 0) {
+exports.category_delete_post = [
+  authorizationValidator,
+  (req, res, next) => {
+    const errors = validationResult(req);
+    Item.find({ category: req.params.id })
+      .then((items) => {
+        if (items.length > 0 || !errors.isEmpty()) {
         // all items not yet deleted
-        Category.findById(req.params.id)
-          .then((category) => {
-            res.render('category_delete', {
-              title: 'Delete Item',
-              category,
-              items,
-            });
+          Category.findById(req.params.id)
+            .then((category) => {
+              res.render('category_delete', {
+                title: 'Delete Item',
+                category,
+                items,
+                errors: errors.array(),
+              });
+            })
+            .catch((err) => next(err));
+          return;
+        }
+        Category.findByIdAndDelete(req.params.id)
+          .then(() => {
+            res.redirect('/inventory');
           })
           .catch((err) => next(err));
-        return;
-      }
-      Category.findByIdAndDelete(req.params.id)
-        .then(() => {
-          res.redirect('/inventory');
-        })
-        .catch((err) => next(err));
-    })
-    .catch((err) => next(err));
-};
+      })
+      .catch((err) => next(err));
+  }];
 
 // display category update form
 exports.category_update_get = (req, res, next) => {
@@ -146,6 +161,7 @@ exports.category_update_get = (req, res, next) => {
       res.render('category_form', {
         title: 'Update Category',
         category,
+        isUpdate: true,
       });
     })
     .catch((err) => next(err));
@@ -153,7 +169,10 @@ exports.category_update_get = (req, res, next) => {
 
 // handle category update form
 exports.category_update_post = [
+  upload.single('uploaded_file'),
   [...categoryFormValidators],
+  authorizationValidator,
+  fileValidator,
   (req, res, next) => {
     const errors = validationResult(req);
     const category = new Category({
@@ -162,11 +181,16 @@ exports.category_update_post = [
       _id: req.params.id,
     });
 
+    if (req.file) {
+      category.image_path = req.file.filename;
+    }
+
     if (!errors.isEmpty()) {
       // has errors
       res.render('category_form', {
         title: 'Update Category',
         category,
+        isUpdate: true,
         errors: errors.array(),
       });
     }
